front: Removed support for Xperia Play device.
