+++ AndroidManifest.xml
@@ -202,20 +202,8 @@
             android:label="@string/GameActivity_title"
             android:theme="@style/appTheme.Black" >
         </activity>
-        <activity
-            android:name="paulscode.android.mupen64plusae.game.GameActivityXperiaPlay"
-            android:configChanges="keyboardHidden|orientation|screenSize"
-            android:label="@string/GameActivity_title"
-            android:theme="@style/appTheme.Black" >
-            <meta-data
-                android:name="android.app.lib_name"
-                android:value="xperia-touchpad" />
-        </activity>
         <service android:name="paulscode.android.mupen64plusae.task.CacheRomInfoService" />
 
-        <meta-data
-            android:name="xperiaplayoptimized_content"
-            android:resource="@drawable/hireslogo" />
         <meta-data
             android:name="game_display_name"
             android:resource="@string/app_name" />
